server/requirements.txt
bidict==0.21.2
gunicorn==20.0.4
python-engineio==4.0.0
python-socketio==5.0.4
toml==0.10.2
